pyproject.toml
[project]
name = "agno-quickstart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.2.16",
    "arxiv>=2.2.0",
    "azure-ai-inference>=1.0.0b9",
    "duckduckgo-search>=8.0.0",
    "fastapi>=0.115.12",
    "googlesearch-python>=1.3.0",
    "openai>=1.72.0",
    "pgvector>=0.4.0",
    "psycopg>=3.2.6",
    "pycountry>=24.6.1",
    "pydantic>=2.11.3",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.0",
    "yfinance>=0.2.55",
]

[project.optional-dependencies]
# Group just for linting with ruff.
lint = [
    "ruff>=0.0.241"
]
dev = [
    "pytest>=7.2",       # For running tests.
    "pytest-cov>=4.0",   # Test coverage reporting.
    "mypy>=1.5",         # Static type checking.
    "ruff>=0.0.241"      # Code linting.
]

# Optional: Additional configuration for ruff.
[tool.ruff]
# Set the maximum line length (this example uses 88 to align with common tools like Black).
line-length = 88
# List of error/warning codes to check. Adjust these to fit your team’s standards.
select = ["A", "ANN", "B", "C", "D", "E", "F", "I", "N", "Q", "R", "S", "T", "W", "UP"]
# Optionally, you can list codes to ignore:
ignore = []

